restructure test-utils folder
