+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
-.env
+.env*
+!.env.dist
 .idea
 xunit.xml
 build/
