--- DEPENDENCIES.md
@@ -1,8 +1,0 @@
-## Dependencies
-These are dependencies needed for packages dappkit makes use of and should only need these for development.
-
-### Python 3
-https://www.python.org/downloads/
-
-### node-gyp
-https://github.com/nodejs/node-gyp#installation
