+++ test/base/web3-connection.spec.ts
@@ -3,7 +3,8 @@ import {expect} from 'chai';
 import {Web3Connection} from '@base/web3-connection';
 import {Web3ConnectionOptions} from '@interfaces/web3-connection-options';
 import {Errors} from '@interfaces/error-enum';
-import {getPrivateKeyFromFile} from '../utils/';
+
+import {getPrivateKeyFromFile} from "../utils/get-pvt-k-from-file";
 
 describe(`Web3Connection`, () => {
   it(`start() fails because missing web3host`, () => {
