+++ test/models/erc4626.spec.ts
@@ -2,7 +2,6 @@ import {describe} from "mocha";
 import {
   defaultWeb3Connection,
   erc20Deployer,
-  getPrivateKeyFromFile,
   hasTxBlockNumber,
   modelExtensionDeployer,
 } from "../utils";
@@ -11,6 +10,7 @@ import {ERC20, Web3Connection} from "../../src";
 import {ERC4626} from "../../src/models/erc4626";
 import {expect} from "chai";
 import {type Web3Account} from "web3-eth-accounts/lib/types";
+import {getPrivateKeyFromFile} from "../utils/get-pvt-k-from-file";
 
 
 describe(`ERC4626`, () => {
