+++ test/models/staking.spec.ts
@@ -3,7 +3,7 @@ import {
   defaultWeb3Connection,
   erc20Deployer, getChainDate,
   hasTxBlockNumber,
-  increaseTime, outputDeploy
+  increaseTime,
 } from '../utils/';
 import {toSmartContractDecimals} from '../../src';
 import {addMinutes, differenceInSeconds} from 'date-fns'
@@ -128,9 +128,5 @@ describe(`StakingContract`, () => {
       expect(withdrawAmount).to.eq(expectedWithdraw);
     });
 
-  });
-
-  after(() => {
-    outputDeploy([[`StakingContract`, contractAddress], [`ERC20`, stakeTokenAddress]])
   });
 })
