+++ test/utils/index.ts
@@ -1,77 +1,12 @@
-import {ERC20, toSmartContractDecimals, Web3Connection, Web3ConnectionOptions} from '../../src';
-import {readFileSync} from 'fs';
-import {resolve} from 'path';
 import {expect, use} from 'chai';
-import Web3 from 'web3';
+
 import {Log} from "../../src/interfaces/web3-core";
 import like from 'chai-like';
 import things from 'chai-things';
-import {type JsonRpcPayload} from "web3-types/src/json_rpc_types";
-import {TransactionReceipt} from "web3-types";
+
 
 use(like);
 use(things);
-
-
-/**
- * Returns the private key from the output provided by ganache
- * @param index
- */
-export function getPrivateKeyFromFile(index = 0) {
-  return Object.values(JSON.parse(readFileSync(resolve('./keys.json'), 'utf-8')).private_keys)[index] as string;
-}
-
-/**
- * A helper function to start a web3connection from process.env files
- * @param start
- * @param revert
- */
-export async function defaultWeb3Connection(start = false, revert = false) {
-  const options: Web3ConnectionOptions = {
-    web3Host: process.env.WEB3_HOST_PROVIDER || 'HTTP://127.0.0.1:8545',
-    privateKey: process.env.WALLET_PRIVATE_KEY || getPrivateKeyFromFile(),
-    skipWindowAssignment: true,
-    restartModelOnDeploy: false,
-  }
-
-  const web3Connection = new Web3Connection(options);
-  if (start)
-    await web3Connection.start();
-  if (revert)
-    await revertChain(web3Connection.Web3);
-
-  return web3Connection;
-}
-
-/**
- * Helper to deploy ERC20 tokens
- * @param name
- * @param symbol
- * @param cap
- * @param web3Connection
- * @returns Promise<TransactionReceipt>
- */
-export async function erc20Deployer(name: string, symbol: string, cap = toSmartContractDecimals(1000000, 18), web3Connection: Web3Connection|Web3ConnectionOptions) {
-  if (!(web3Connection instanceof Web3Connection))
-    web3Connection = new Web3Connection(web3Connection)
-
-  await web3Connection.start();
-
-  const deployer = new ERC20(web3Connection);
-  await deployer.loadAbi();
-
-  const address = await deployer.connection.getAddress();
-  return deployer.deployJsonAbi(name, symbol, cap, address);
-}
-
-/**
- * Create a new account and return it
- * @param web3Connection
- * @param salt
- */
-export function newWeb3Account(web3Connection: Web3Connection) {
-  return web3Connection.Web3.eth.accounts.create();
-}
 
 /**
  * Helper to assert that the promise needs to be rejected
@@ -87,48 +22,6 @@ export async function shouldBeRejected(promise: Promise<any>, withErrorMessage?:
       expect((e?.innerError?.message || e?.message)).to.contain(withErrorMessage);
     else expect(e).to.exist;
   }
-}
-
-const payload = (method: string, params: any[] = []) => ({jsonrpc: `2.0`, method, params, id: 0});
-
-/**
- * Advance in time by mining a block on the chain
- * @param {number} time seconds to increase
- * @param {Web3} web3 web3
- */
-export async function increaseTime(time: number, web3: Web3) {
-
-  const timeAdvance = payload(`evm_increaseTime`, [time]) as JsonRpcPayload;
-  const mine = payload(`evm_mine`, []) as JsonRpcPayload;
-  const provider = web3.currentProvider!;
-
-  return new Promise((resolve, reject) => {
-    provider.send(timeAdvance, (err,) => {
-      if (err)
-        reject(err);
-      else provider.send(mine, (err, resp) => {
-        if (err)
-          reject(err)
-        resolve(resp);
-      })
-    })
-  });
-}
-
-/**
- * Revert the chain to its default state
- * @param web3
- */
-export async function revertChain(web3: Web3) {
-  return new Promise((resolve, reject) => {
-    web3.currentProvider!
-      .send(payload(`evm_revert`, []) as JsonRpcPayload,
-            (err, resp) => {
-              if (err)
-                reject(err)
-              resolve(resp);
-            })
-  })
 }
 
 /**
@@ -147,45 +40,23 @@ export async function hasTxBlockNumber(promise: Promise<any>, message = `Should 
     return tx;
 }
 
-export async function expectEvent<T = any>(promise: Promise<T>, eventName: string, params: any) {
+/**
+ * Assert that the transaction from the promivent has the needed decoded args and event name
+ */
+export async function expectEvent<T = any>(promise: Promise<T>, eventName: string, params?: {[k: string]: any}) {
   const tx = await hasTxBlockNumber(promise);
   expect(tx.logs, `Should have logs array`).to.exist;
 
   expect(tx.logs).to.be.an('array').that.contains.something.like({event: eventName});
-  expect(tx.logs.map((l: Log) => l.args)).to.be.an('array').that.contains.something.like(params);
+
+  if (params)
+    expect(tx.logs.map((l: Log) => l.args)).to.be.an('array').that.contains.something.like(params);
 }
 
-export function calculateAPR(apr = 1, start = 0,
-                                   end = 0, amount = 1) {
-  const timePassed = +end-+start;
-  const ms = 60;
-  return ((((apr / 365 / 24 / ms) * timePassed) / ms) * amount) / 100;
-}
-
-/**
- * Return chain date
- * @param web3Connection
- * @returns Date
- */
-export async function getChainDate(web3Connection: Web3Connection) {
-  return new Date(parseInt((await web3Connection.eth.getBlock(await web3Connection.Web3.eth.getBlockNumber())).timestamp.toString(),10) * 1000)
-}
-
-export function outputDeploy(info: [string, string][] = []) {
-  if (!process.env.DEBUG_TESTS)
-    return;
-
-  console.log(`Deployed`, info.map(([name, address]) => `\n\t${name}:\t${address}`).join(``))
-}
-
-/**
- * Agnostic model deployer to help out in the deployment of initial contracts
- * @param web3connection
- * @param _class
- * @param args
- */
-export function modelExtensionDeployer(web3connection: Web3Connection, _class: any, args?: any[]): Promise<TransactionReceipt> {
-  const deployer = new _class(web3connection);
-  deployer.loadAbi();
-  return deployer.deployJsonAbi(...(args || []));
-}
+export {getPrivateKeyFromFile} from "./get-pvt-k-from-file";
+export {defaultWeb3Connection} from "./default-web3-connection";
+export {erc20Deployer} from "./erc20-deployer";
+export {increaseTime} from "./increase-chain-time";
+export {revertChain} from "./revert-chain";
+export {getChainDate} from "./get-chain-date";
+export {modelExtensionDeployer} from "./model-deployer";
